Add jekyll-include-cache as a dependency
After more than a year of deploying this with any changes, my build is
failing.  It may be due to weird version errors with Ruby (my system
seems to be using 2.7.1, where the current version is 3.1), however I
found a similar error online which states that jekyll-include-cache is
needed.  Trying to add this and hope it fixes the error.

<https://github.com/mmistakes/minimal-mistakes/issues/1875>